Refactor stop deletion confirmation dialog to its own component
